Refactor control components: add Toggle, Switch, Matrix, and SlideShow; remove old implementations
