+++ src/lib/components/controls/SlideShow.svelte
