+++ src/lib/components/controls/Slider.svelte
