+++ src/lib/components/controls/index.ts
@@ -1,0 +1,8 @@
+import Dropdown from './Dropdown.svelte';
+import Matrix from './Matrix.svelte';
+import Slider from './Slider.svelte';
+import SlideShow from './SlideShow.svelte';
+import Switch from './Switch.svelte';
+import Toggle from './Toggle.svelte';
+
+export { Dropdown, Matrix, SlideShow, Slider, Switch, Toggle };
